Refactor: 컴포넌트 위치 및 이름 수정
