+++ frontend/src/components/cards/Common/index.tsx
