+++ frontend/src/components/cards/Common/style.ts
